[#26] refactor: applying DDD-principle oriented package structure
