+++ BE/src/main/java/dev/codesquad/airbnb02/domain/host/business/HostService.java
@@ -1,0 +1,5 @@
+package dev.codesquad.airbnb02.domain.host.business;
+
+public class HostService {
+
+}
