+++ BE/src/main/java/dev/codesquad/airbnb02/domain/host/data/HostRepository.java
@@ -1,6 +1,6 @@
-package dev.codesquad.airbnb02.repository;
+package dev.codesquad.airbnb02.domain.host.data;
 
-import dev.codesquad.airbnb02.entity.Host;
+import dev.codesquad.airbnb02.domain.host.domain.Host;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface HostRepository extends JpaRepository<Host,Long> {
