+++ BE/src/main/java/dev/codesquad/airbnb02/domain/host/domain/Host.java
@@ -1,7 +1,8 @@
-package dev.codesquad.airbnb02.entity;
+package dev.codesquad.airbnb02.domain.host.domain;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.sun.istack.NotNull;
+import dev.codesquad.airbnb02.domain.room.domain.Room;
 import java.util.List;
 import javax.persistence.OneToMany;
 import lombok.Getter;
