+++ BE/src/main/java/dev/codesquad/airbnb02/domain/room/business/RoomService.java
@@ -1,0 +1,5 @@
+package dev.codesquad.airbnb02.domain.room.business;
+
+public class RoomService {
+
+}
