+++ BE/src/main/java/dev/codesquad/airbnb02/domain/room/data/RoomRepository.java
@@ -1,6 +1,6 @@
-package dev.codesquad.airbnb02.repository;
+package dev.codesquad.airbnb02.domain.room.data;
 
-import dev.codesquad.airbnb02.entity.Room;
+import dev.codesquad.airbnb02.domain.room.domain.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface RoomRepository extends JpaRepository<Room,Long> {
